+++ tests/test_configuration.py
@@ -371,6 +371,21 @@ class TestConfiguration(unittest.TestCase):
             assert len(record) == 1
             assert c.send_code is False
 
+    def test_validate_send_environment(self):
+        c = Configuration()
+        assert c.send_environment is True
+        with pytest.warns(RuntimeWarning) as record:
+            c.configure(send_environment='False')
+
+            assert len(record) == 1
+            assert (str(record[0].message) ==
+                    'send_environment should be bool, got str')
+            assert c.send_environment is True
+
+            c.configure(send_environment=False)
+            assert len(record) == 1
+            assert c.send_environment is False
+
     def test_validate_session_endpoint(self):
         c = Configuration()
         with pytest.warns(RuntimeWarning) as record: